refactor: add ModalContainer and ButtonContainer
