+++ src/components/common/ButtonContainer.js
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles';
+
+const ButtonContainer = styled('div')(() => ({
+    display: 'flex',
+    justifyContent: 'flex-end',
+  }));
+
+  export default ButtonContainer;
